backend/requirements.txt
asgiref==3.3.4
Django==3.2.3
django-cors-headers==3.7.0
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
Pillow==8.4.0
PyJWT==2.3.0
python-decouple==3.4
pytz==2021.1
sqlparse==0.4.1
